Split DummyRng from the RSA key implementation
Move DummyRng to the separate module to allow it to be used from
PaddingScheme module.
